\name{lmem-package}
\alias{lmem-package}
\alias{lmem}
\docType{package}
\title{
\packageTitle{lmem}
}
\description{
\packageDescription{lmem}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{lmem}
\packageIndices{lmem}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{lmem}

Maintainer: \packageMaintainer{lmem}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
